+++ src/modules/ratings/ratings.module.ts
@@ -3,7 +3,7 @@ import { Module } from '@nestjs/common'
 import { RatingsRepository } from './ratings.repository'
 import { Note } from '../notes/models/notes.model'
 import { SequelizeModule } from '@nestjs/sequelize'
-import { DB_CONNECTION_NAME } from 'src/common/constants'
+import { DB_CONNECTION_NAME } from '../../common/constants'
 import { Rating } from './models/rating.model'
 import { User } from '../class/models/user.model'