+++ src/modules/ratings/ratings.repository.ts
@@ -1,6 +1,6 @@
 import { BadRequestException, Injectable } from '@nestjs/common'
 import { InjectModel } from '@nestjs/sequelize'
-import { DB_CONNECTION_NAME } from 'src/common/constants'
+import { DB_CONNECTION_NAME } from '../../common/constants'
 import { User } from '../class/models/user.model'
 import { Note } from '../notes/models/notes.model'
 import { Rating } from './models/rating.model'